dn_detection_app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
onnxruntime==1.16.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
SQLAlchemy==2.0.23
sqlite3
pytest==7.4.3
httpx==0.25.2
gunicorn==21.2.0
python-dotenv==1.0.0